+++ spec/dummy/src/config/configuration.example.js
@@ -48,6 +48,7 @@ export default new Configuration({
       }
     }
   },
+  cookieSecret: "dummy-cookie-secret",
   directory: dummyDirectory(),
   environment: "development",
   environmentHandler: new NodeEnvironmentHandler(),