+++ spec/dummy/src/config/configuration.peakflow.pgsql.js
@@ -55,6 +55,7 @@ export default new Configuration({
       }
     }
   },
+  cookieSecret: "dummy-cookie-secret",
   directory: dummyDirectory(),
   environmentHandler: new NodeEnvironmentHandler(),
   initializeModels: async ({configuration}) => {